feat: stream mode

+++ router.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const router = require('koa-router')()
-const { writeFile, execCommand, languageList } = require('./util/index')
+const { writeFile, execCommand, languageList, spawnCommand } = require('./util/index')
 const SSE = require('./util/sse')
 
 const hostPath = path.join(__dirname, '/code')
@@ -10,28 +10,48 @@ let idx = 0
 router.get('/', ctx => ctx.render('index', { languageList: ['node', 'python', 'go'] }))
 
 router.post('/runner', async (ctx) => {
-  const { language, code, version, uid } = ctx.request.body
+  const { language, code, version, uid, streamMode } = ctx.request.body
   const { extension, dockerPrefix, command } = languageList[language]
   const filename = `main-${++idx}${extension}`
 
   SSE.writeStream(uid, 'sse-message', 'sand box: 开始拉取镜像...')
-  await Promise.all([
+  const [[writeErr, ], [pullErr, ]] = await Promise.all([
     writeFile(hostPath, filename, code, () => {
       SSE.writeStream(uid, 'sse-message', 'sand box: 文件写入异常')
     }),
-    execCommand(`docker pull ${dockerPrefix}:${version}`, { timeout: 30000 }, () => {
-      SSE.writeStream(uid, 'sse-message', 'sand box: 镜像拉取超时')
+    execCommand(`docker pull ${dockerPrefix}:${version}`, {
+      handleTimeout: () => {
+        SSE.writeStream(uid, 'sse-message', 'sand box: 镜像拉取超时')
+      },
+      option: {
+        timeout: 30000
+      }
     })
   ])
   SSE.writeStream(uid, 'sse-message', 'sand box: 镜像拉取成功')
 
-  const [e, data] = await execCommand(
-    `docker run --rm --memory=50m --name runner-${idx} -v ${hostPath}:${workPath} ${dockerPrefix}:${version} ${command} /code/${filename}`,
-    { timeout: 10000 },
-    `docker stop runner-${idx}`
-  )
-  SSE.writeStream(uid, 'sse-result', { result: e || data } )
-  SSE.writeStream(uid, 'sse-message', 'sand box: 执行完毕')
+  if (!writeErr || !pullErr) {
+    const commandStr = `docker run --rm --memory=50m --name runner-${idx} -v ${hostPath}:${workPath} ${dockerPrefix}:${version} ${command} /code/${filename}`
+    const exitCommand = `docker stop runner-${idx}`
+
+    if (!streamMode) {
+      const [e, data] = await execCommand(commandStr, {
+        handleTimeout: exitCommand,
+        option: {
+          timeout: 10000
+        }
+      })
+      SSE.writeStream(uid, 'sse-result', { result: e || data } )
+      SSE.writeStream(uid, 'sse-message', 'sand box: 执行完毕')
+    } else {
+      spawnCommand(commandStr, SSE.getInstance(uid).stream, {
+        onExit: () => {
+          SSE.writeStream(uid, 'sse-message', 'sand box: 执行完毕')
+          execCommand(exitCommand)
+        }
+      })
+    }
+  }
   return ctx.body = { }
 })
